Handle piped input to STDIN properly.
Piping input from a cURL command yields unexepected results.
On most systems (e.g. MacOS), the pipe command gives a bytestream
that is buffered (64 KB on my system), which prevents trancription
to operate properly :
https://unix.stackexchange.com/questions/11946/how-big-is-the-pipe-buffer

One way to fix is to get more data (fs.readSync) when necessary
and fill the audio buffer we send to Google, therefore changing
'setInterval' to a recursive 'setTimeout' call that offers more
flexibility in setting the timer.