EOF condition when data is no longer available.
If we want to be able to work in a real time fashion, we need to
consider that the audio file we're processing is kept fed by an
external process. Therefore we don't want to rely on the file size
to determine the EOF condition.

If we get 0 bytes of data from the audio files after 5 consecutive
iterations, then we have an End Of File condition and we exit.